Trim fclShopMisc FUN_003f7730 parameters

+++ src/Event/mt_evtMain.c
@@ -638,7 +638,7 @@ void FUN_003650c0(u32 param_1,u64 param_2,int param_3);
 void FUN_00365250(u32 param_1,u64 param_2,int param_3);
 void FUN_00365360(u32 param_1,u64 param_2,int param_3);
 void FUN_00365430(int param_1,int param_2);
-void FUN_003657e0(int param_1,u64 param_2);
+void FUN_003657e0(int param_1,int param_2);
 void FUN_003658b0(u32 param_1,int param_2);
 void FUN_00365a40(u32 param_1,int param_2);
 void FUN_00365b70(u32 param_1,u64 param_2);
@@ -5694,7 +5694,7 @@ u16 * FUN_003655f0(u32 param_1,int param_2)
 // FUN_003657E0 NONMATCHING
 
 
-void FUN_003657e0(int param_1,u64 param_2)
+void FUN_003657e0(int param_1,int param_2)
